Move ApiMemberExtractor to a top-level class
